[front] - fix: correct import of datadogLogger to match file rename
 - Rename datadogLogger import statements to align with the updated filename
 - Ensure consistent and accurate references across different components in the agent builder directory
