+++ front/components/agent_builder/capabilities/mcp/utils/formStateUtils.ts
@@ -4,7 +4,7 @@ import type { MCPFormData } from "@app/components/agent_builder/AgentBuilderForm
 import { getDefaultFormValues } from "@app/components/agent_builder/capabilities/mcp/utils/formDefaults";
 import type { AgentBuilderAction } from "@app/components/agent_builder/types";
 import type { MCPServerViewType } from "@app/lib/api/mcp";
-import datadogLogger from "@app/logger/datadogsLogger";
+import datadogLogger from "@app/logger/datadogLogger";
 
 /**
  * Creates form reset function for different dialog modes
